tests/wpt/web-platform-tests/html/syntax/parsing-html-fragments/the-input-byte-stream-001.html

<!DOCTYPE html>
<html  lang="en" >
<head>
  <title>HTTP charset</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='https://html.spec.whatwg.org/multipage/#the-input-byte-stream'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name='flags' content='http'>
<style type='text/css'>
.test div { width: 50px; }</style>
<link rel="stylesheet" type="text/css" href="support/encodingtests-15.css">
</head>
<body>



<div class='test'><div id='box' class='ýäè'>&#xA0;</div></div>


<!--Notes:

The test contains a div with a class name that contains the following sequence of bytes: 0xC3 0xBD 0xC3 0xA4 0xC3 0xA8. These represent different sequences of characters in ISO 8859-15, ISO 8859-1 and UTF-8. The external, UTF-8-encoded stylesheet contains a selector <code>.test div.&#x00C3;&#x0153;&#x00C3;&#x20AC;&#x00C3;&#x0161;</code>. This matches the sequence of bytes above when they are interpreted as ISO 8859-15. If the class name matches the selector then the test will pass.

The only character encoding declaration for this HTML file is in the HTTP header, which sets the encoding to ISO 8859-15.

-->
<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 100);
}, "The character encoding of a page can be set using the HTTP header charset declaration.");
</script>

<div id='log'></div>

</body>
</html>
